+++ src/harvest/task.cr
@@ -3,7 +3,7 @@ module Harvest
   class Task
     include JSON::Serializable
 
-    property id : Int32
+    property id : Int64
     property name : String
     property billable_by_default : Bool
     property created_at : Time
